% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{df_labelled_to_factor}
\alias{df_labelled_to_factor}
\title{Takes a data frame containing variables of class \code{haven::labelled} and replaces them with factors}
\usage{
df_labelled_to_factor(ldf)
}
\arguments{
\item{ldf}{A data frame which may contained variables of class \code{haven::labelled}}
}
\value{
A data frame in which labeled variables have been replaced with factors
}
\description{
Takes a data frame containing variables of class \code{haven::labelled} and replaces them with factors
}
\examples{
fp <- system.file("extdata", "survey_data.sav", package = "zmarket")
orig <- haven::read_sav(fp)
rep <- df_labelled_to_factor(orig)
lapply(orig, "class")
lapply(rep, "class")
}
