% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_data}
\alias{experiment_data}
\title{Responses to survey experiment about messages that affect likelihood of download}
\format{
A data frame with 10,704 rows and 9 variables:
}
\source{
\url{https://www.gradientmetrics.com}
}
\usage{
experiment_data
}
\description{
A dataset from a Gradient Metrics survey experiment designed to determine the effects of attributes of a message on the motivation to download.
}
\keyword{datasets}
