% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Calculate the index of a variable}
\usage{
index(grouping_variable, index_variable)
}
\arguments{
\item{grouping_variable}{A discrete grouping variable}

\item{index_variable}{A discrete variable for which to calculate index}
}
\value{
A table of index values which are equal to 100 times the frequency of a level of \code{index_variable} within levels of \code{grouping_variable} relative to the aggregate of all its levels.
}
\description{
Calculate the index of a variable
}
\examples{
with(experiment_data, index(duration, social_proof))
with(survey_data, index(s_gender, s_age))
plot(with(survey_data, index(s_gender, s_age)))
}
